Final submit first version
